+++ src/service.rs
@@ -54,7 +54,7 @@ struct QueryRoot {
 
 #[derive(SimpleObject)]
 struct GameState {
-    game_id: u32,
+    game_id: u16,
     board: [[u16; 4]; 4],
     is_ended: bool,
     score: u64,
@@ -62,7 +62,7 @@ struct GameState {
 
 #[Object]
 impl QueryRoot {
-    async fn game(&self, game_id: u32) -> Option<GameState> {
+    async fn game(&self, game_id: u16) -> Option<GameState> {
         if let Ok(Some(game)) = self.state.games.try_load_entry(&game_id).await {
             let game_state = GameState {
                 game_id: *game.game_id.get(),
@@ -81,12 +81,12 @@ struct MutationRoot;
 
 #[Object]
 impl MutationRoot {
-    async fn new_game(&self, seed: Option<u32>) -> Vec<u8> {
+    async fn new_game(&self, seed: Option<u16>) -> Vec<u8> {
         let seed = seed.unwrap_or(0);
         bcs::to_bytes(&Operation::NewGame { seed }).unwrap()
     }
 
-    async fn make_move(&self, game_id: u32, direction: Direction) -> Vec<u8> {
+    async fn make_move(&self, game_id: u16, direction: Direction) -> Vec<u8> {
         let operation = Operation::MakeMove { game_id, direction };
         bcs::to_bytes(&operation).unwrap()
     }
